XRDDEV-900 Refactor backup filename validation
- Move filename validation logic from repository class to utils class.
- Add filename validation to all repository methods that read, create or delete backup files.

+++ src/proxy-ui-api/src/main/java/org/niis/xroad/restapi/repository/BackupRepository.java
@@ -27,6 +27,7 @@ package org.niis.xroad.restapi.repository;
 import ee.ria.xroad.common.SystemProperties;
 
 import lombok.extern.slf4j.Slf4j;
+import org.niis.xroad.restapi.util.FormatUtils;
 import org.springframework.stereotype.Repository;
 
 import java.io.File;
@@ -38,7 +39,6 @@ import java.nio.file.attribute.BasicFileAttributes;
 import java.time.OffsetDateTime;
 import java.time.ZoneOffset;
 import java.util.List;
-import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -50,9 +50,6 @@ import java.util.stream.Stream;
 public class BackupRepository {
 
     private static final String CONFIGURATION_BACKUP_PATH = SystemProperties.getConfBackupPath();
-    // Criteria for valid backup file names:
-    // 1) cannot start with ".", 2) must contain one or more word characters ([a-zA-Z_0-9.-]), 3) must end with ".tar"
-    private static final String BACKUP_FILENAME_PATTERN = "^(?!\\.)[\\w\\.\\-]+\\.tar$";
     // Set maximum number of levels of directories to visit, subdirectories are excluded
     private static final int DIR_MAX_DEPTH = 1;
 
@@ -62,7 +59,7 @@ public class BackupRepository {
      */
     public List<File> getBackupFiles() {
         try (Stream<Path> walk = Files.walk(Paths.get(CONFIGURATION_BACKUP_PATH), DIR_MAX_DEPTH)) {
-            return walk.filter(this::isFilenameValid).map(this::getFile).collect(Collectors.toList());
+            return walk.filter(this::isValidBackupFilename).map(this::getFile).collect(Collectors.toList());
         } catch (IOException ioe) {
             log.error("can't read backup files from configuration path (" + CONFIGURATION_BACKUP_PATH + ")");
             throw new RuntimeException(ioe);
@@ -90,6 +87,9 @@ public class BackupRepository {
      * @param filename
      */
     public void deleteBackupFile(String filename) {
+        if (!FormatUtils.isValidBackupFilename(filename)) {
+            throw new RuntimeException("invalid backup filename");
+        }
         Path path = getFilePath(filename);
         try {
             Files.deleteIfExists(path);
@@ -105,6 +105,9 @@ public class BackupRepository {
      * @return
      */
     public byte[] readBackupFile(String filename) {
+        if (!FormatUtils.isValidBackupFilename(filename)) {
+            throw new RuntimeException("invalid backup filename");
+        }
         Path path = getFilePath(filename);
         try {
             return Files.readAllBytes(path);
@@ -121,10 +124,10 @@ public class BackupRepository {
      * @return
      */
     public OffsetDateTime writeBackupFile(String filename, byte[] content) {
-        Path path = getFilePath(filename);
-        if (!isFilenameValid(path)) {
+        if (!FormatUtils.isValidBackupFilename(filename)) {
             throw new RuntimeException("invalid backup filename");
         }
+        Path path = getFilePath(filename);
         try {
             Files.write(path, content);
             return getCreatedAt(path.getFileName().toString());
@@ -151,20 +154,11 @@ public class BackupRepository {
 
     /**
      * Check if the given filename is valid and meets the defined criteria
-     * @param filename
-     * @return
-     */
-    public boolean isFilenameValid(String filename) {
-        return isFilenameValid(getFilePath(filename));
-    }
-
-    /**
-     * Check if the given filename is valid and meets the defined criteria
      * @param path
      * @return
      */
-    public boolean isFilenameValid(Path path) {
-        return Pattern.compile(BACKUP_FILENAME_PATTERN).matcher(path.getFileName().toString()).matches();
+    private boolean isValidBackupFilename(Path path) {
+        return FormatUtils.isValidBackupFilename(path.getFileName().toString());
     }
 
     /**
